# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: WaCalls API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    HTTP API for WaCalls, a self-hosted WhatsApp calling bridge. All /api/*
    endpoints require the bearer token configured via WACALLS_API_TOKEN
    (when unset, authentication is disabled). The SSE stream also accepts the
    token as the access_token query parameter. This document is the canonical
    contract: a test gate fails the build when routes and spec drift apart.
servers:
  - url: /
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/sessions:
    get:
      operationId: listSessions
      summary: List WhatsApp accounts
      responses:
        "200":
          description: All configured sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionInfo"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      operationId: createSession
      summary: Create an account and begin QR pairing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display name, defaults to "Session"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}:
    delete:
      operationId: deleteSession
      summary: Log out and remove an account
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Session removed
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    patch:
      operationId: renameSession
      summary: Rename an account
      parameters:
        - $ref: "#/components/parameters/Sid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: New display name
      responses:
        "204":
          description: Session renamed
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/logout:
    post:
      operationId: logoutSession
      summary: Disconnect an account, keeping it for re-pairing
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Session disconnected
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/pair:
    post:
      operationId: pairSession
      summary: Re-pair an account, emitting a fresh QR over SSE
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Pairing restarted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/onwhatsapp:
    post:
      operationId: isOnWhatsApp
      summary: Check which of the given phone numbers are registered on WhatsApp
      description: >-
        Added by the fork. Dialling a number without WhatsApp burns an attempt,
        raises the chip's failure rate and worsens how the account looks to Meta,
        so a list importer should check before queueing.


        Returns one result per number asked, in the order they were sent. The
        underlying USync query only answers about the users the server knows, so
        aligning results by position would silently put one contact's answer on
        another's row.


        Numbers may be sent with or without the leading `+`. At most 50 per
        request: the query goes to Meta's servers and volume draws attention, so
        an importer is expected to space its batches out and spread them across
        the tenant's chips.
      parameters:
        - $ref: "#/components/parameters/Sid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phones]
              properties:
                phones:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: { type: string }
              example:
                phones: ["5514981120008", "5514996035887"]
      responses:
        "200":
          description: One result per number asked, in the order they were sent
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [phone, exists]
                      properties:
                        phone:
                          type: string
                          description: The number exactly as it was sent
                        exists: { type: boolean }
                        jid:
                          type: string
                          description: Canonical JID, absent when the number is not on WhatsApp
              example:
                results:
                  - phone: "5514981120008"
                    exists: true
                    jid: "5514981120008@s.whatsapp.net"
                  - phone: "5514996035887"
                    exists: false
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The session has no WhatsApp account linked yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "502":
          description: WhatsApp did not answer the query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/sessions/{sid}/calls:
    post:
      operationId: startCall
      summary: Start an outgoing call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: E.164-ish phone number; non-digits are stripped
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Operator identified by X-Client-Id is already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Per-session concurrent call cap reached, or rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Session is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listCalls
      summary: Live calls of the session, same shape as the SSE call-list
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Live calls ordered by startedAt then callId
          content:
            application/json:
              schema:
                type: object
                required: [calls]
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}:
    get:
      operationId: getCall
      summary: One live call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: The live call; 404 once it ends or if it belongs to another session
          content:
            application/json:
              schema:
                type: object
                required: [call]
                properties:
                  call:
                    $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      operationId: endCall
      summary: End an active call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: Call ended (idempotent)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/webrtc:
    post:
      operationId: exchangeWebRTC
      summary: Exchange the browser WebRTC SDP for the call audio bridge
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sdp_offer]
              properties:
                sdp_offer:
                  type: string
      responses:
        "200":
          description: SDP answer
          content:
            application/json:
              schema:
                type: object
                required: [sdp_answer]
                properties:
                  sdp_answer:
                    type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      operationId: acceptCall
      summary: Accept an incoming call, claiming it for the X-Client-Id operator
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Claimed by another client, or operator already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      operationId: rejectCall
      summary: Reject an incoming call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: Rejected (idempotent when the call is already gone)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is in a state that cannot be rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/mute:
    post:
      operationId: setCallMute
      summary: Set the local microphone mute state, signaling it to the peer
      description: >-
        Signals mute_v2 to the peer so their client reflects the mute state. The
        audio itself is silenced by the media source (the browser mic track); this
        endpoint only carries the signaling and call state.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [muted]
              properties:
                muted:
                  type: boolean
      responses:
        "200":
          description: Mute state applied and signaled
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "400":
          description: Missing or invalid muted field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/play:
    post:
      operationId: playAnnouncement
      summary: Play a pre-recorded announcement into a live call
      description: >-
        Pushes an audio asset onto the call as if it came from the operator, paced
        one 20 ms frame at a time. While it plays, audio arriving from the browser
        leg is dropped rather than mixed, so the announcement cannot be talked
        over. The announcement is included in the recording.


        The returned duration tells the caller when the line is free. Completion
        is also published on the event stream as call-announce-done, which is the
        reliable signal because playback can be cut short.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset]
              properties:
                asset:
                  type: string
                  description: >-
                    Asset name in the audio directory, without the .wav
                    extension. Must match ^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$.
                  example: aviso-tenant-abc
                interruptible:
                  type: boolean
                  description: >-
                    Barge-in is not implemented. Only false is accepted; true is
                    rejected rather than silently ignored.
                  default: false
      responses:
        "200":
          description: Playback started
          content:
            application/json:
              schema:
                type: object
                required: [durationMs]
                properties:
                  durationMs:
                    type: integer
                    format: int64
                    description: How long the asset takes to play.
                    example: 4620
        "400":
          description: Missing asset, unsupported flag, or an asset that is not mono 16 kHz PCM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such session, call, or asset, or announcements are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/stopplay:
    post:
      operationId: stopAnnouncement
      summary: Stop the announcement playing on a call
      description: >-
        Idempotent. Returns success whether or not an announcement was playing,
        because teardown reaches this from more than one path.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: Nothing is playing on this call any more
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/recording:
    get:
      operationId: getRecording
      summary: Download the recording of a finished call
      description: >-
        Streams the stereo WAV written during the call: left channel is what was
        sent to the contact (operator microphone plus any announcement), right
        channel is what the contact sent. PCM signed 16-bit little endian, two
        channels, 16 kHz.


        Only available once the call has ended, because the two size fields in the
        WAV header are written on close. Range requests are supported, so an
        interrupted download can resume.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: The recording
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        "400":
          description: Session or call id contains characters that are not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No recording for this call, or recording is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The call is still in progress, so the file is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteRecording
      summary: Delete a recording from local disk
      description: >-
        Called after the recording has been stored elsewhere. Idempotent: a
        recording that is already gone counts as deleted. Refused while the call
        is still in progress, because unlinking a file the recorder still holds
        open loses the recording without any error surfacing.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: The recording is no longer on disk
        "400":
          description: Session or call id contains characters that are not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Recording is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The call is still in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/audio/{name}:
    put:
      operationId: putAudioAsset
      summary: Upload an announcement asset
      description: >-
        Stores an audio asset on the engine's disk so it can be played into calls.
        The engine holds no storage credentials of its own, so the caller pushes
        each asset here once. Keeping assets local also keeps the network off the
        call setup path, where a slow fetch would mean a caller hearing silence.


        The body is the raw WAV file, which must be mono, 16 kHz, 16-bit PCM.
        Anything else is rejected rather than resampled. Re-uploading the same
        name replaces the asset and the cached copy. Requests are capped at 1 MiB,
        which is around 32 seconds of audio.
      parameters:
        - name: name
          in: path
          required: true
          description: Asset name, without extension. Must match ^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$.
          schema:
            type: string
          example: aviso-tenant-abc
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Asset stored
          content:
            application/json:
              schema:
                type: object
                required: [durationMs]
                properties:
                  durationMs:
                    type: integer
                    format: int64
                    example: 4620
        "400":
          description: Invalid name, or a file that is not mono 16 kHz 16-bit PCM WAV
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Announcements are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Body larger than 1 MiB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteAudioAsset
      summary: Delete an announcement asset
      description: Idempotent. An asset that is already gone counts as deleted.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The asset is no longer available
        "400":
          description: Invalid asset name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Announcements are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/history:
    get:
      operationId: listHistory
      summary: Ended calls, newest first, keyset-paginated
      parameters:
        - $ref: "#/components/parameters/Sid"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
          description: Page size; values above 200 are clamped to 200
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor from the previous page's nextCursor
      responses:
        "200":
          description: One page of ended calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/history/export:
    get:
      operationId: exportHistory
      summary: Full call history as CSV
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: >-
            CSV with header callId,direction,peer,owner,startedAt,endedAt,endReason;
            timestamps are RFC 3339 UTC and owner is empty when unclaimed
          content:
            text/csv:
              schema:
                type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/login:
    post:
      operationId: login
      summary: Admin login; sets an httpOnly session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Logged in; sets the wacalls_session cookie
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/logout:
    post:
      operationId: logout
      summary: Invalidate the current session and clear the cookie
      responses:
        "204":
          description: Logged out
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    queryToken:
      type: apiKey
      in: query
      name: access_token
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      schema:
        type: string
      description: Session id
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such session or call
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request parameter or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SessionInfo:
      type: object
      required: [id, name, jid, state, paired]
      properties:
        id:
          type: string
        name:
          type: string
        jid:
          type: string
        state:
          type: string
        paired:
          type: boolean
        photoUrl:
          type: string
          description: Profile picture URL of the connected WhatsApp account, when available
...